Fix swiftlint violations: extract SourceCodeDetector into separate file
Move source code detection logic (containsSourceCode, looksLikeCode)
from Clipboard into a dedicated SourceCodeDetector struct to stay
within swiftlint file_length and type_body_length limits.
